+++ src/PointLight.cpp
@@ -1,12 +1,11 @@
 #include "PointLight.hpp"
 
 
-PointLight::PointLight(const Vector3f& position, const Vector3f& color) :
+PointLight::PointLight(const Vector3f& position, const Vector3d& color) :
     position_(position), color_(color)
 {}
 
 LightSample PointLight::drawSample(const Vector3f& p) {
-    //std::lock_guard<std::mutex> lock(mutex_);
     Vector3f pd = p-position_;
 
     LightSample s;
